+++ tests/test_consul_discovery.py
@@ -54,15 +54,17 @@ class AgentBasedTests(helpers.EnvironmentMixin, unittest.TestCase):
         for service_id in list(self._service_ids):
             self.deregister_service(service_id, ignore_error=True)
 
-    def register_service(self):
+    def register_service(self, meta=None, port=None):
         service_id = str(uuid.uuid4())
         service_details = {
             'ID': service_id,
             'Name': 's' + service_id.replace('-', '').lower(),
             'Address': '10.0.0.1',
-            'Port': random.randint(10000, 20000),
+            'Port': port or random.randint(10000, 20000),
             'Datacenter': self.datacenter,
         }
+        if meta:
+            service_details['Meta'] = meta
         response = self.session.put(self.agent_url + '/service/register',
                                     json=service_details)
         response.raise_for_status()
@@ -98,3 +100,18 @@ class AgentBasedTests(helpers.EnvironmentMixin, unittest.TestCase):
         self.deregister_service(service_info['ID'])
         self.assertEqual(expected,
                          klempner.url.build_url(service_info['Name']))
+
+    def test_that_scheme_set_by_metadata(self):
+        service_info = self.register_service(meta={'protocol': 'https'})
+        expected = 'https://{Name}.service.{Datacenter}.consul:{Port}/'.format(
+            **service_info)
+        self.assertEqual(expected,
+                         klempner.url.build_url(service_info['Name']))
+
+    def test_that_scheme_set_by_port(self):
+        for port, scheme in klempner.config.URL_SCHEME_MAP.items():
+            service_info = self.register_service(port=port)
+            self.assertEqual(
+                '{scheme}://{Name}.service.{Datacenter}.consul:{port}/'.format(
+                    port=port, scheme=scheme, **service_info),
+                klempner.url.build_url(service_info['Name']))